refactor: 내비게이션에 학습스크린 추가 및 화면 이동 가능하도록 수정
